Attempt to better support fallback executable paths on Windows
